// src/types/global.d.ts
export interface Career {
  name: string;
  icon: React.ElementType;
  desc: string;
}

export interface Goal {
  goal: string;
  icon: React.ElementType;
  desc: string;
}

export interface Level {
  level: string;
  icon: React.ElementType;
  desc: string;
}

export interface Mentor {
  name: string;
  imageUrl: string;
  attribute: string;
  desc: string;
}

export interface Duration {
  months: string;
}